// front-end/src/app/models/organizer.ts
export class Organizer {
  _id?: string;
  name: string;
  address_web: string;
  start_date: number;
  end_date: number;
  logo: string;
  publicity_image: string;
  limit_speaker_time: number;
  politics: string;
  capacity_speakers: number;
  capacity_participants: number;
  status_congress: boolean;
}